Updated GUI program
GUI aspect of the project updated and organized. The GUI now divided into 3 main files: main.py, gui.py, and events.py.

- main.py is the designated main file for the GUI and project as a whole (serving as the project's main function), and handles any and all events seen within the program
- gui.py is tasked strictly with formatting and generating the GUI and the GUI elements
- events.py stores the code that results from when an event is triggered (used to visually separate event triggers and event actions) 

Going forward, files and functions from the audio branch of the project can be easily and seamlessly imported into GUI files (primarily main.py and events.py) to add features on the audio branch. In summary, functions from the audio branch will be able to be triggered with GUI events.

Necessary GUI elements to properly display and implement project's features are still required, the files uploaded in this change specifically are a proof of concept of how external functions can be run by the GUI.